+++ pkgs/google_cloud_pubsub/lib/google_cloud_pubsub.dart
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 export 'package:google_cloud_rpc/exceptions.dart';
+export 'package:google_cloud_rpc/retry.dart' show ExponentialRetry, RetryRunner;
 export 'package:grpc/grpc.dart'
     show
         BaseAuthenticator,
@@ -23,6 +24,6 @@ export 'package:grpc/grpc.dart'
 export 'src/batching.dart' show BatchingSettings;
 export 'src/client.dart' show PubSub;
 export 'src/message.dart' show Message, ReceivedMessage;
-export 'src/retry.dart' show RetrySettings;
+export 'src/retry.dart' show defaultPubSubRetry;
 export 'src/subscription.dart' show AckSettings, Subscription;
 export 'src/topic.dart' show PublishSettings, Topic;